refactor(position): update import paths for consistency and clarity
