+++ src/modules/position/dtos/create-position.dto.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 
-import { PositionStatus } from '../enums';
+import { PositionStatus } from '@/modules/position/enums';
 
 export const createPositionDto = z.object({
   name: z.string(),
